Isha | Completed other two pages.
