+++ src/components/NewComps/Banner/SplitBanner/index.js
@@ -2,7 +2,7 @@ import BookFreeDemoForm from "../../BookFreeDemoForm";
 import { trustBadgeBrands } from "../../TrustBadge";
 import style from "./SplitBanner.module.scss";
 
-export default function SplitBanner() {
+export default function SplitBanner({ compdata }) {
   return (
     <>
       <section className={`py-3 mb-5 outfit-font ${style.banner}`}>
@@ -11,7 +11,6 @@ export default function SplitBanner() {
             <img
               src="/img/giddh-logo.svg"
               alt="Giddh Logo"
-              className="banner-logo"
               width={120}
               height={24}
             />
@@ -29,18 +28,15 @@ export default function SplitBanner() {
           </div>
           <div className="row align-items-center justify-content-between garmond-font">
             <div className="col-lg-6 col-md-6 col-12 mb-4 mb-lg-0 pe-3 d-flex flex-column gap-2">
-              <span className="c-fs-4 outfit-font">
+              <span className="font-md outfit-font">
                 {" "}
-                BUILD FOR <span className="col-primary"> SMBs</span>
+                BUILD FOR <span className="font-primary"> SMBs</span>
               </span>
-              <h1 className="font-heading col-primary">
-                Smart, Simple & All-in-One Accounting for Growing Businesses
+              <h1 className="font-heading font-primary">
+                {compdata?.heading}
               </h1>
-              <p className="col-grey outfit-font ">
-                Giddh brings everything together on a secure, cloud platform
-                built for growing businesses. Simplifies accounting with
-                GST-ready billing, smart inventory sync, and real-time financial
-                insights.
+              <p className="font-grey outfit-font">
+                {compdata?.subHeading}
               </p>
               <div className="d-flex flex-wrap align-items-center gap-4">
                 {trustBadgeBrands?.map((logo, index) => (
@@ -48,7 +44,6 @@ export default function SplitBanner() {
                     <img
                       src={logo.src}
                       alt={logo.alt}
-                      className="logo-item"
                       height={34}
                     />
                   </a>
