+++ src/components/NewComps/CTASection/CTASection.module.scss
@@ -1,0 +1,13 @@
+.inner {
+  grid-template-columns: 1fr 480px;
+
+  @media (max-width: 992px) {
+    grid-template-columns: 1fr;
+  }
+}
+
+@media screen and (max-width: 992px) {
+  .imageWrap {
+    display: none;
+  }
+}
