+++ src/components/NewComps/Features/ToggleFeatures/index.js
@@ -1,14 +1,12 @@
-import data from "./data.json";
 import style from "./ToggleFeatures.module.scss";
 import { useState, useCallback, memo } from "react";
 import Image from "next/image";
 import * as MdIcons from "react-icons/md";
 import { MdCircle } from "react-icons/md";
 
-export default function ToggleFeatures() {
+export default function ToggleFeatures({ features }) {
   const [expandedFeature, setExpandedFeature] = useState(0);
-  const Icon = MdIcons[data[expandedFeature].icon];
-
+  const Icon = MdIcons[features[expandedFeature].icon];
   const handleFeatureToggle = useCallback((index) => {
     setExpandedFeature((prev) => (prev === index ? prev : index));
   }, []);
@@ -17,14 +15,14 @@ export default function ToggleFeatures() {
     <section className="py-5">
       <div className="container">
         <h2 className="new-sub-heading text-center garmond-font  mb-3 mb-md-4">
-        <span className="col-primary">Giddh Features</span> That Make Business Easy
+          <span className="col-primary">Giddh Features</span> That Make Business Easy
         </h2>
 
         <div className="d-flex flex-column flex-sm-row gap-4 outfit-font align-items-center justify-content-between">
           <div className="col-12 col-sm-6 order-1 order-md-1">
             <div id="features" className="p-2">
-              {data?.length > 0 &&
-                data.map((feature, index) => (
+              {features?.length > 0 &&
+                features.map((feature, index) => (
                   <FeatureItem
                     key={index}
                     feature={feature}
@@ -40,8 +38,8 @@ export default function ToggleFeatures() {
             <div className="d-flex align-items-center justify-content-center h-100 ">
               <div className="position-relative d-inline-block rounded overflow-hidden ">
                 <Image
-                  src={data[expandedFeature].image}
-                  alt={data[expandedFeature].name}
+                  src={features[expandedFeature].image}
+                  alt={features[expandedFeature].name}
                   height={400}
                   width={400}
                   className={`cursor-pointer ${style.image}`}
@@ -52,10 +50,10 @@ export default function ToggleFeatures() {
                 >
                   <div className="text-start text-white">
                     <h3 className="font-outfit text-white c-fw-600 c-fs-4 text-start">
-                      {data[expandedFeature].name}
+                      {features[expandedFeature].name}
                     </h3>
                     <p className="c-fs-5 text-start">
-                      {data[expandedFeature].description}
+                      {features[expandedFeature].description}
                     </p>
                   </div>
                   <button
@@ -65,7 +63,7 @@ export default function ToggleFeatures() {
                       {Icon && <Icon size={18} className="text-white" />}
                     </div>
                     <span className="text-white">
-                      {data[expandedFeature].name}
+                      {features[expandedFeature].name}
                     </span>
                   </button>
                 </div>
@@ -82,32 +80,28 @@ const FeatureItem = memo(({ feature, index, isExpanded, onToggle }) => {
   const handleToggle = useCallback(() => {
     onToggle(index);
   }, [index, onToggle]);
-
   const Icon = MdIcons[feature.icon];
 
   return (
     <div
-      className={`d-flex flex-column gap-1 rounded p-2 ${
-        isExpanded ? "bg-accent text-white border" : ""
-      } cursor-pointer`}
+      className={`d-flex flex-column gap-1 rounded p-2 ${isExpanded ? "bg-accent text-white border" : ""
+        } cursor-pointer`}
     >
       <div className="d-flex align-items-center gap-2" onClick={handleToggle}>
         <div className="border border-col-primary p-2 bg-light-blue rounded d-flex align-items-center justify-content-center">
           {Icon && <Icon size={18} className="col-primary" />}
         </div>
         <h3
-          className={`col-primary c-fw-600 c-fs-4 m-0 ${
-            isExpanded ? "text-white" : ""
-          }`}
+          className={`col-primary c-fw-600 c-fs-4 m-0 ${isExpanded ? "text-white" : ""
+            }`}
         >
           {feature?.name}
         </h3>
       </div>
 
       <div
-        className={`d-flex flex-column gap-2 pt-2 ps-sm-5 ${
-          isExpanded ? "d-flex" : "d-none"
-        }`}
+        className={`d-flex flex-column gap-2 pt-2 ps-sm-5 ${isExpanded ? "d-flex" : "d-none"
+          }`}
       >
         <div className="d-block d-sm-none">
           <Image
