--- src/components/NewComps/GiddhFor/data.json
@@ -1,34 +1,0 @@
-{
-  "heading": "Who Can Grow with Giddh?",
-  "subheading": "Giddh provides your business with clarity, control, and hassle-free compliance, making it ideal for product sellers, service providers, and multi-client operations.",
-  "content": [
-    {
-      "title":"Retailers",
-      "name": "Retailers & Traders",
-      "description": "Get real-time control of sales, inventory, and finances for smooth growth.",
-      "icon": "🛒",
-      "image": "/img/online-accounting-software/Retailers.svg"
-    },
-    {
-      "title":"Startups",
-      "name": "Startups & Service Providers",
-      "description": "Manage cash flow, track expenses, and scale smarter with real-time financial insights.",
-      "icon": "🚀",
-      "image": "/img/online-accounting-software/Startups.svg"
-    },
-    {
-      "title":"Accountants",
-      "name": "Accountants & Finance Teams",
-      "description": " Handle multiple clients, automate reports, and ensure accurate compliance with ease.",
-      "icon": "📊",
-      "image": "/img/online-accounting-software/Accountants.svg"
-    },
-    {
-      "title":"E-commerce",
-      "name": "Shopify & E-commerce Brands",
-      "description": "Sync sales, manage stock, and track payments seamlessly across all online platforms.",
-      "icon": "🛍️",
-      "image": "/img/online-accounting-software/E-commerce.svg"
-    }
-  ]
-}
