+++ src/components/NewComps/Stats/index.js
@@ -2,7 +2,7 @@ export default function Stats({ stats }) {
   return (
     <section className="py-5 outfit-font bg-accent">
       <div className="container d-flex flex-column gap-4">
-        <h2 className="c-fw-400 garmond-font sub-heading text-center text-white">
+        <h2 className="font-400 garmond-font font-sub-heading text-center font-white">
           {stats?.heading}
         </h2>
         <div className="row justify-content-center g-2">
@@ -10,10 +10,10 @@ export default function Stats({ stats }) {
             <div key={index} className="col-12 col-sm-6 col-md-4 d-flex">
               <div
                 className="w-100 d-flex flex-column align-items-center justify-content-center
-                           border border-white rounded-2 py-4 px-3 h-100 text-center"
+                           border border-white rounded-2 py-4 px-3 h-100 text-center font-white"
               >
-                <div className="text-white fw-bold c-fs-1">{stat?.value}</div>
-                <div className="text-white fs-6">{stat?.label}</div>
+                <div className="font-600 font-xxl">{stat?.value}</div>
+                <div className="font-sm">{stat?.label}</div>
               </div>
             </div>
           ))}
