--- src/components/NewComps/Testimonials/data.json
@@ -1,72 +1,0 @@
-[
-  {
-    "company": "Sinead North",
-    "owner": "Christian Schiel",
-    "testimonial": "Giddh is a great solution to manage customers and invoices internationally. The multi-currency option makes it very convenient, and generating invoices is simple and quick. I can also easily manage VAT reports in the UK. The support team is patient, understands the challenges we face, and resolves them effectively."
-  },
-  {
-    "company": "Dvara e-registry",
-    "owner": "Priyanshi Priya",
-    "testimonial": "Giddh is an excellent platform for managing multiple companies seamlessly. It allows me to handle different businesses in real time, all in one place. Switching between companies is incredibly fast and effortless, making the entire process smooth and efficient."
-  },
-  {
-    "company": "Teknoworth Medical Care OPC Pvt Ltd",
-    "owner": "Imran",
-    "testimonial": "Giddh is the most reliable application for inventory management in both India and the UAE. Managing data is effortless, and the mobile app works flawlessly, ensuring smooth operations anytime, anywhere."
-  },
-  {
-    "company": "Shruti Corporation",
-    "owner": "Shruti",
-    "testimonial": "Giddh is one of the easiest ledger-based accounting systems I’ve used. Creating ledger entries is quick and effortless, making it truly ready to go. Even someone without a strong accounting background can easily manage and run their business with it."
-  },
-  {
-    "company": "PI Green Consulting Pvt Ltd",
-    "owner": null,
-    "testimonial": "The application works seamlessly, allowing multiple users to access it in real time without any limits or lags. Multi-login functionality is effortless, making collaboration smooth and efficient."
-  },
-  {
-    "company": "Synergene Global",
-    "owner": "Manish Badonia",
-    "testimonial": "The support team is excellent, and the latest version of Giddh comes packed with powerful features. Giddh can be accessed from anywhere with just an internet connection, making it highly convenient and reliable."
-  },
-  {
-    "company": "Attic India",
-    "owner": "Kuldeep",
-    "testimonial": "The inventory reports work brilliantly. Creating parent and child groups is simple and effective, allowing the inventory to be organized and bifurcated properly for better clarity."
-  },
-  {
-    "company": "KARUNA TEMP SYSTEM PRIVATE LIMITED",
-    "owner": "Rohan Pavithran",
-    "testimonial": "Giddh’s estimates and proforma invoices work seamlessly. We can create proformas and easily convert them into invoices, tracking their status as paid, unpaid, or partially paid. Automated WhatsApp messages and email notifications after creating vouchers significantly reduce manual effort."
-  },
-  {
-    "company": "Dental Designs",
-    "owner": "Dr Gins Paul",
-    "testimonial": "Giddh offers multiple types of invoices, including cash and sales invoices, making billing flexible and convenient. The dashboard provides a clear summary of the business’s health on a single screen, helping us instantly understand where we stand on any given day."
-  },
-  {
-    "company": "Taurus Lifestyle",
-    "owner": "Kunal",
-    "testimonial": "The inventory module in Giddh is extremely powerful. Its variant creation feature allows multiple variants for a single item, which is very helpful for detailed inventory management. The barcode scanning feature further completes the requirements, making the process fast and efficient."
-  },
-  {
-    "company": "M/S STUDIO47",
-    "owner": null,
-    "testimonial": "Giddh provides a wide range of reports including financial statements, daybook, sales and purchase registers, inventory reports, and GST reports. Having all these reports in one place makes managing business insights simple and efficient."
-  },
-  {
-    "company": "SP Wellness",
-    "owner": null,
-    "testimonial": "Support connects positively when needed. The ledger accounting module makes it easy to create entries; I can easily manage my expenses by creating the accounts and making entries there."
-  },
-  {
-    "company": "The Medicity",
-    "owner": null,
-    "testimonial": "With Giddh, managing bank accounts is simple and hassle-free. Incomes and expenses can be tracked directly, and the built-in GST filing option makes compliance much easier. It’s truly a great all-in-one solution."
-  },
-  {
-    "company": "Mitti [Organic. Fresh. Desi]",
-    "owner": "Ankita Chadha",
-    "testimonial": "The inventory management helps in restaurant management and the branch creation helps to manage multiple branches under a single platform. It helps to see the branch profiles along with the warehouses."
-  }
-]
