--- src/components/NewComps/WhyGiddh/data.json
@@ -1,37 +1,0 @@
-[
-    {
-        "features": "Multi-Company Access",
-        "other": false,
-        "giddh": true
-    },
-    {
-        "features": "Unlimited User Access",
-        "other": false,
-        "giddh": true
-    },
-    {
-        "features": "Ledger-Based Accounting",
-        "other": false,
-        "giddh": true
-    },
-    {
-        "features": "White-Label Option",
-        "other": false,
-        "giddh": true
-    },
-    {
-        "features": "GST Compliance",
-        "other": true,
-        "giddh": true
-    },
-    {
-        "features": "Multi-Currency Accounting",
-        "other": false,
-        "giddh": true
-    },
-    {
-        "features": "Inventory Management",
-        "other": true,
-        "giddh": true
-    }
-]
