+++ src/pages/index.js
@@ -1,5 +1,5 @@
 import Home from "@/components/Pages/Home";
 
-export default function Index({ pageData, pageInfo}) {
+export default function Index({ pageData, pageInfo }) {
   return <Home pageData={pageData} pageInfo={pageInfo} />;
 }
