+++ src/scss/_mixins.scss
@@ -52,6 +52,7 @@ $color-values: (
   light-golden: --col-light-golden,
   extra-light-blue: --col-extra-light-blue,
   slate-grey: --col-slate-grey,
+  accent-light: --col-accent-light,
 );
 
 // Generate background color classes using universal mixin
