+++ src/scss/_text.scss
@@ -153,11 +153,11 @@
   font-size: var(--font-sub-heading);
 }
 
-// font weight
+// new font weight
 
-.fontw-400 {
+.font-400 {
   font-weight: var(--c-fw-400) !important;
 }
-.fontw-600 {
+.font-600 {
   font-weight: var(--c-fw-600) !important;
 }
