+++ src/utils/getPageData.js
@@ -1,6 +1,5 @@
 export default function getPageData(pageInfo) {
-  const pagesToGetData = ["home" , "cloud-accounting-software"];
-
+  const pagesToGetData = ["home", "cloud-accounting-software", "invoice-software", "multi-user-accounting-software" , "online-accounting-software"];
   if (!pagesToGetData.includes(pageInfo?.page)) return null;
   try {
     const pageData = require(`@/data/pagesdata/${pageInfo?.country}/${pageInfo?.page}.json`);
